LocationObservableListBuilder V1
I created an object class that will build a masterlist of location
objects then parse those locations into three observable lists. The
lists are sorted alphbetically and "All" will always be the first item
in the list.
The constructor is "public LocationObservableListBuilder()"
the constructors for other methods you can use are:
public void addToAllLocationsList(String camp,String bldg, String rm)
this is used to added full locations to the master list
public void buildLists() this should be called after all locations are
added from the database
public void reBuildLists(String campSelection, String bldgSelection,
String rmSelection) this method should be called each time the user
selects an option from the dropdown fields in the GUI or if you need to
clear and rebuild the observable lists
public void printObservableLists() is callable if you want to print the
current Observable lists to the console